Optimize code, and merge unbatch_edge_index into unbatch.py
