+++ torch_geometric/utils/__init__.py
@@ -17,8 +17,7 @@ from .mask import index_to_mask, mask_to_index
 from .to_dense_batch import to_dense_batch
 from .to_dense_adj import to_dense_adj
 from .sparse import dense_to_sparse
-from .unbatch import unbatch
-from .unbatch_edge_index import unbatch_edge_index
+from .unbatch import unbatch, unbatch_edge_index
 from .normalized_cut import normalized_cut
 from .grid import grid
 from .geodesic import geodesic_distance
