+++ torch_geometric/utils/unbatch.py
@@ -22,3 +22,27 @@ def unbatch(src: Tensor, batch: Tensor, dim: int = 0) -> List[Tensor]:
     """
     sizes = degree(batch, dtype=torch.long).tolist()
     return src.split(sizes, dim)
+
+
+def unbatch_edge_index(edge_index: Tensor, batch: Tensor) -> List[Tensor]:
+    r"""Splits the :obj:`edge_index` according to a :obj:`batch` vector.
+
+    Args:
+        edge_index (Tensor): The edge_index tensor. Must be sorted.
+        batch (LongTensor): The batch vector
+            :math:`\mathbf{b} \in {\{ 0, \ldots, B-1\}}^N`, which assigns each
+            node to a specific example. Must be ordered, consecutive, and
+            starts with 0.
+
+    :rtype: :class:`List[Tensor]`
+    """
+    boundary = torch.cumsum(degree(batch), dim=0)
+    inc = torch.cat([boundary.new_tensor([0]), boundary[:-1]], dim=0)
+
+    edge_batch = batch[edge_index[0]]
+    sizes = degree(edge_batch).long().cpu().tolist()
+
+    out = [(edge - inc[batch_idx]).to(torch.int64)
+           for batch_idx, edge in enumerate(edge_index.split(sizes, dim=1))]
+
+    return out
